Added the people page

+++ src/backend/Modules/Users/Modules.Users.Application/Followers/GetFollowerStats/FollowerStatsResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace Modules.Users.Application.Followers.GetFollowerStats;
 
-public sealed record FollowerStatsResponse(Guid UserId, int FollowerCount, int FollowingCount);
+public sealed record FollowerStatsResponse(Guid UserId, long FollowerCount, long FollowingCount);
